+++ libs/portals/my-pages/health/src/screens/HealthConversations/utils/messagingWindow.ts
@@ -1,10 +1,16 @@
-import {
-  formatMinutes,
-  parseTimeToMinutes,
-} from '@island.is/portals/my-pages/core'
-
 const MINUTES_IN_DAY = 24 * 60
 const CLOSING_SOON_THRESHOLD_MINUTES = 30
+
+const pad = (value: number) => value.toString().padStart(2, '0')
+
+const formatMinutes = (minutes: number) =>
+  `${pad(Math.floor(minutes / 60))}:${pad(minutes % 60)}`
+
+const parseTimeToMinutes = (time: string): number | undefined => {
+  const match = time.match(/^(\d{1,2}):(\d{2})/)
+  if (!match) return undefined
+  return Number(match[1]) * 60 + Number(match[2])
+}
 
 
 export const getMessagingWindowInfo = ({
